+++ packages/interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # interfaces
 
+## 0.1.1
+
+### Patch Changes
+
+- Release changeset type
+
 ## 0.1.0
 
 ### Minor Changes